Create webscrapingGoogleSearch.py
scraping Google search results using Python:
First  function is to grabs the HTML from a Google.com search results page. 
Then parsing the HTML, will allow  to extract the elements we want from the Google results page. 
For this i will be using BeautifulSoup, this library makes it very easily to extract the data  from a webpage.
There are a number of different errors that could be thrown , it is essential  to catch all of these possible exceptions.